+++ src/meshes/remove_cells.jl
