+++ src/models/submodels/boundary_conditions.jl
@@ -1,0 +1,4 @@
+import Jutul: JutulEntity
+export BoundaryDirichletFaces
+
+struct BoundaryDirichletFaces <: JutulEntity end
