+++ src/models/submodels/current_collector.jl
