+++ test/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+BattMo = "6f0c0536-3c2c-4762-a987-c605a8a6f898"
 Jutul = "2b460a1a-8a2b-45b2-b125-b5c536396eb9"
 MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
